+++ userspace/libsanalyzer/src/app_check_emitter.cpp
@@ -3,7 +3,7 @@
 #include "configuration_manager.h"
 
 app_check_emitter::app_check_emitter(app_checks_proxy::metric_map_t& app_metrics,
-				     const uint16_t app_metrics_limit,
+				     const unsigned int app_metrics_limit,
 				     const prometheus_conf& prom_conf,
 				     std::unordered_map<std::string, std::tuple<unsigned, unsigned>>& app_checks_by_container,
 				     std::unordered_map<std::string, std::tuple<unsigned, unsigned>>& prometheus_by_container,
